+++ src/pydiverse/pipedag/backend/__init__.py
@@ -1,3 +1,4 @@
 from .blob import *
 from .lock import *
 from .table import *
+from .table_cache import *